move ContextVisitor to separate file
